+++ flexiapi/app/Http/Middleware/IsSpaceExpired.php
@@ -11,15 +11,15 @@ class IsSpaceExpired
 {
     public function handle(Request $request, Closure $next): Response
     {
-        if (empty(config('app.root_domain'))) {
-            return abort(503, 'APP_ROOT_DOMAIN is not configured');
+        if (empty(config('app.root_host'))) {
+            return abort(503, 'APP_ROOT_HOST is not configured');
         }
 
         $space = \App\Space::where('host', $request->header('host'))->first();
 
         if ($space) {
-            if (!str_ends_with($space->host, config('app.root_domain'))) {
-                return abort(503, 'The APP_ROOT_DOMAIN configured does not match with the current root domain');
+            if (!str_ends_with($space->host, config('app.root_host'))) {
+                return abort(503, 'The APP_ROOT_HOST configured does not match with the current root domain');
             }
 
             Config::set('app.url', '://' . $space->host);