+++ flexiapi/config/app.php
@@ -14,7 +14,7 @@ return [
     */
 
     'name' => env('APP_NAME', 'Account Manager'),
-    'sip_domain' => env('APP_SIP_DOMAIN', 'sip.domain.com'),
+    //'sip_domain' => 'sip.domain.com',
 
     'project_url' => env('APP_PROJECT_URL', ''),
     'terms_of_use_url' => env('TERMS_OF_USE_URL', ''),
@@ -128,7 +128,7 @@ return [
     */
 
     'url' => env('APP_URL', 'http://localhost'),
-    'root_domain' => env('APP_ROOT_DOMAIN', null),
+    'root_host' => env('APP_ROOT_HOST', null),
     'asset_url' => env('ASSET_URL', null),
 
     /*