+++ CHANGELOG.md
@@ -7,6 +7,7 @@ v1.7
 - Fix GH-15 Add password import from CSV
 - Fix FLEXIAPI-242 Add stricter validation for the AccountCreationToken Push Notification endpoint
 - Fix FLEXIAPI-241 Add a /push-notification endpoint to send custom push notifications to the Flexisip Pusher
+- Fix FLEXIAPI-250 Allow Spaces to be declared without a subdomain
 
 v1.6
 ----